Refactor track handlers
Separated functions in different files.
Reduce code repetition.
Add config files accordingly.
