+++ src/shared/ui/Carousel/config/FlowAxis.js
@@ -1,0 +1,2 @@
+export const HORIZONTAL = "X";
+export const VERTICAL = "Y";
