+++ src/shared/ui/Carousel/config/FlowDirection.js
@@ -1,0 +1,2 @@
+export const POSITIVE = "+";
+export const NEGATIVE = "-";
